+++ packages/astro/package.json
@@ -157,6 +157,7 @@
     "chai": "^4.3.6",
     "cheerio": "^1.0.0-rc.10",
     "mocha": "^9.2.2",
+    "readable-stream": "^3.6.0",
     "sass": "^1.49.9"
   },
   "engines": {
