+++ packages/astro/src/vite-plugin-astro/styles.ts
@@ -1,6 +1,6 @@
 import type * as vite from 'vite';
 
-import { STYLE_EXTENSIONS } from '../core/render/dev/css.js';
+import { STYLE_EXTENSIONS } from '../core/render/util.js';
 
 export type TransformHook = (code: string, id: string, ssr?: boolean) => Promise<vite.TransformResult>;
