+++ packages/integrations/deno/CHANGELOG.md
@@ -1,0 +1,7 @@
+# @astrojs/node
+
+## 0.0.2-next.0
+
+### Patch Changes
+
+- [#2873](https://github.com/withastro/astro/pull/2873) [`e4025d1f`](https://github.com/withastro/astro/commit/e4025d1f530310d6ab951109f4f53878a307471a) Thanks [@matthewp](https://github.com/matthewp)! - Improves the build by building to a single file for rendering
