+++ packages/integrations/deno/src/shim.ts
@@ -1,0 +1,5 @@
+
+(globalThis as any).process = {
+	argv: [],
+	env: {}
+};
